Magit: reference window will checkout remote branches with RET
Reference window (y r) will checkout local and remote branches, as well
as tags (with a detached head). Checkout on remote branches just creates
a local branch with the same name pointing at the remote branch that
originated it.

This restores the behaviour of RET as it was on magit 1.x