Move spacemacs-useless-buffers-regexp setting
This variable setting has always been inside dotspacemacs/init()
function, and it has recently stopped working i.e. magit buffers are
still treated as useless at startup.

This moves the variable setting to dotspacemacs/user-config. Thinking
about it now, it seems to be the proper place for it, since this
variable name does not start with "dotspacemacs".